requirements.txt
boto3>=1.15.6
pyftpdlib>=1.5.6
PyYAML>=5.1.2
